refactor: separate mention functionality from comment input component (#5277)
* feat(authenticator): initialize task management template on sign-in


* refactor: separate mention functionality from comment input component


---------
